AST

Description

License : MIT

Documentation

fromAST

Arguments

:: Lexer2SynAna
Input graphASTforest.

-> String
Portable text representation of the AST:

Each function is represented by its own section. A section has a header and content; it continues either until the next section, a blank line or the end of the file, whichever comes first.

A section header consists of a single line containing the name of the function, enclosed in square brackets, e. g. [function_name]. There cannot be any whitespace before the opening bracket.

The section content consists of zero or more non-blank lines containing exactly three records delimited by a semicolon ;. Each line describes a node and contains the following records, in this order:

The node ID (numeric), e. g. 1. * The Rail lexeme, e. g. o or [constant] etc. Note that track lexemes like - or + are not included in the graph. Multi-character lexemes like constants may include semicolons, so you need to parse them correctly! In other words, you need to take care of lines like 1;[some ; constant];2. * Node ID of the follower node, e. g. 2. May be 0 if there is no next node.

Convert a graph/AST into a portable text representation. See also fromGraph.

toAST

Arguments

:: String
Portable text representation. See fromAST.

-> Lexer2SynAna
Output graph.

Convert a portable text representation of a graph into a concrete graph representation. See also toGraph. See fromAST for a specification of the portable text representation.

fromGraph

Arguments

:: Graph
Input graph.

-> String
Text representation.

Convert an Graph for a single function to a portable text representation. See fromAST for a specification of the representation.

TODO: Currently, this apparently crashes the program on invalid input. More sensible error handling? At least a nice error message would be nice.

splitfunctions

Arguments

:: String
Portable text representation of the forest.

-> [[String]]
List of lists, each being a list of lines making up a separate function graph.

Split a portable text representation of multiple function graphs (a forest) into separate text representations of each function graph.

toGraph

Arguments

:: [String]
List of lines making up the text representation of the function.

-> Graph
Graph describing the function.

Convert a portable text representation of a single function into an Graph. Raises errors on invalid input (see ErrorHandling).

offset

Arguments

:: Int
Amount to shift node by.

-> LexNode
Node to operate on.

-> LexNode
Shifted node.

Shift a node by the given amount. May be positive or negative. This is used by toGraph and fromGraph to shift all nodes by 1 or -1, respectively, which is done because the portable text representation of the graph does not include a leading Start node with ID 1 -- instead, the node with ID 1 is the first real graph node. In other words, when exporting to the text representation, the Start node is removed and all other nodes are shifted by -1 using this function. When importing, a Start node is added and all nodes are shifted by 1.

adjacent

Arguments

:: Grid2D
Line representation of the current function.

-> IP
Current instruction pointer.

-> (Char, Char, Char)
Adjacent (left secondary, primary, right secondary) symbols

Get adjacent (left secondary, primary, right secondary) symbols for the current IP position.

parse

Arguments

:: Grid2D
Line representation of current function.

-> IP
Current instruction pointer.

-> (Maybe Lexeme, IP)
Resulting lexeme (if any) and the new instruction pointer.

Get the next lexeme at the current position.

stepwhile

Arguments

:: Grid2D
Line representation of current function.

-> IP
Current instruction pointer.

-> (Char -> Bool)
Function: Should return True if collection should stop. Gets the current Char as an argument.

-> (String, IP)
Collected characters and the new instruction pointer.

Collect characters until a condition is met while moving in the current direction.

moveable

Arguments

:: Grid2D
Line representation of current function

-> IP
Current instruction pointer

-> RelDirection
Where to move to

-> Bool
Whether or not the move could be made

Checks if the instruction pointer can be moved without leaving the grid

readconstant

Arguments

:: Grid2D
Current function in line representation

-> IP
Current instruction pointer

-> Char
Opening string delimiter, e. g. '['

-> Char
Closing string delimiter, e. g. ']'

-> (String, IP)
The processed constant and the new instruction pointer

Read a string constant and handle escape sequences like n. Raises an error on invalid escape sequences and badly formatted constants.

escapestring :: String -> String

lookahead

Arguments

:: Grid2D
Line representation of current function

-> IP
Current instruction pointer

-> Int
How many characters of lookahead to produce?

-> String
n characters of lookahead

Lookahead n characters in the current direction.

skip

Arguments

:: Grid2D
Line representation of current function

-> IP
Current instruction pointer

-> Int
How many characters to skip? If 1, this is the same as doing move ip Forward.

-> IP
New instruction pointer

Skip n characters in the current direction and return the new IP.

next :: Grid2D -> IP -> Char

Get the Char at the next position of the instruction pointer

valids

Arguments

:: Grid2D
Line representation of current function.

-> IP
Current instruction pointer.

-> (String, String, String)
Tuple consisting of:

Valid characters for movement to the (relative) left. * Valid characters for movement in the (relative) forward direction. * Valid characters for movement to the (relative) right.

Return valid chars for movement depending on the current direction.

commandchars :: String

turnblocked :: String