ErrorHandling

Description

License : MIT

Contains error messages for errors during the compilation of a rail program that result in an immediate stop of the process. The purpose is to get an overview of all possible error messages for negative test cases for integration testing.

Documentation

generalError :: [Char]

noStartSymbolFound :: [Char]

strFunctionNameMissing :: [Char]

strNestedOpenBracket :: [Char]

strNonSymmetricEscape :: [Char]

strUnhandledEscape :: [Char]

strMissingClosingBracket :: [Char]

strInvalidVarName :: [Char]

strInvalidFuncName :: [Char]

shrLineNoLexeme :: [Char]

strInvalidMovement :: [Char]

strMainMissing :: [Char]

strUnknownNode :: [Char]

strEmptyProgram :: [Char]

strDuplicateFunctions :: [Char]