InstructionPointer

Description

License : MIT

Documentation

data Direction

An absolute direction.

Constructors

N

NE

E

SE

S

SW

W

NW

data RelDirection

A relative direction.

Constructors

Left

Forward

Right

data IP

Instruction pointer consisting of position and an orientation.

Constructors

IP

Fields

count :: Int

Number of processed characters since start of current function.

posx :: Int

Current X position.

posy :: Int

Current Y position.

dir :: Direction

Current Direction.

path :: RelDirection

Determines if the instruction pointer is on a left or right path of a Junction

known :: Map (Int, Int, Direction) Int

Map of Position and Direction to Id

start :: IP

Initial value for the instruction pointer at the start of a function.

crash :: IP

An instruction pointer representing a crash (fatal error).

crashfrom :: IP -> IP

ipmerge :: IP -> IP -> IP

turn :: Char -> IP -> IP

move

Arguments

:: IP
Current instruction pointer.

-> RelDirection
Relative direction to move in.

-> IP
New instruction pointer.

Move the instruction pointer in a relative direction.

current

Arguments

:: Grid2D
Line representation of the current function.

-> IP
Current instruction pointer.

-> Char
Char at the current IP position.

Get the Char at the current position of the instruction pointer.

charat

Arguments

:: Grid2D
Line representation of current function.

-> (Int, Int)
Position as (x, y) coordinate.

-> Char
Char at given position.

Returns Char at given position, ' ' if position is invalid.

junctionturns :: Grid2D -> IP -> (IP, IP)

lambdadirs :: IP -> (IP, IP)

tuplemove :: (IP, IP) -> (IP, IP)

nodecount :: IP -> Int

nodeadd :: IP -> Int -> IP

addpath :: (IP, IP) -> (IP, IP)

turnaround :: IP -> IP

visited

Arguments

:: IP
Instruction pointer to use.

-> Int
ID of visited node or 0 if none.

Get ID of the node that has been already visited using the current IP (direction and coordinates).

posdir

Arguments

:: IP
Current instruction pointer.

-> RelDirection
Current relative direction.

-> (Int, Int)
New position that results from the given relative movement.

Get the position of a specific heading.

posabsdir

Arguments

:: IP
Current instruction pointer.

-> Direction
Current absolute direction.

-> (Int, Int)
New position that results from the given absolute movement.

Get the position of an absolute direction.

absolute

Arguments

:: IP
Current instruction pointer.

-> RelDirection
Relative direction to convert.

-> Direction
Equivalent absolute direction.

Convert a relative direction into a relative one.

primary :: IP -> Char