InterfaceDT

Description

License : MIT

Defining algebraic data types for all compiler stages. Each module in the pipeline has two corresponding algebraic data types, one defining the input and the other defining the output. The output data type of a compiler stage is the input data type of the following compiler stage. The algebraic data types ensure a clean interface between the modules.

Documentation

type Grid2D = Map Int (Map Int Char)

type PositionedGrid = (Grid2D, Int)

type LexNode = (Int, Lexeme, Int)

(NodeID (start: 1), Lexeme of Node, NodeID of following Node (0 if none))

type Graph = (String, [LexNode])

(FunctionID, Graph of Function as adjacency list)

type AST = (String, [(Int, [Lexeme], Int)])

(FunctionID, [(PathID (start: 1), List of Lexemes to be executed sequentially, PathID of following Path)])

data Lexeme

following Nodes or Pathes could have ID==0, in this case there is no Follower

Junction Int = if false goto Int; if true = following node |Lambda Int = Int holds node of anonymous function

Constructors

NOP

Boom

EOF

Input

Output

Underflow

RType

Constant String

Push String

Pop String

Call String

Add1

Divide

Multiply

Remainder

Subtract

Cut

Append

Size

Nil

Cons

Breakup

Greater

Equal

Start

Finish

Junction Int

Lambda Int

data Input2PreProc

Constructors

IIP String

data PreProc2Lexer

Constructors

IPL [PositionedGrid]

data Lexer2SynAna

Constructors

ILS [Graph]

data SynAna2SemAna

Constructors

ISS [AST]

data SemAna2InterCode

Constructors

ISI [AST]

data InterCode2Backend

Constructors

IIB Module

data Backend2Output

Constructors

IBO (IO String)