IntermediateCode

Description

License : MIT Stability : unstable

IntemediateCode.hs takes the output from the SemanticalAnalysis module (which is a list of paths) and generates LLVM IR code. It turns every path of the form (PathID, [Lexeme], PathID) into a basic block.

Documentation

data CodegenState

Constructors

CodegenState

Fields

blocks :: [BasicBlock]

count :: Word

localDict :: Map String (Int, Integer)

newtype Codegen a

Constructors

Codegen

Fields

runCodegen :: State CodegenState a

data GlobalCodegenState

Constructors

GlobalCodegenState

Fields

dict :: Map String (Int, Integer)

newtype GlobalCodegen a

Constructors

GlobalCodegen

Fields

runGlobalCodegen :: State GlobalCodegenState a

execGlobalCodegen :: Map String (Int, Integer) -> GlobalCodegen a -> a

execCodegen :: Map String (Int, Integer) -> Codegen a -> a

generateModule :: [Definition] -> Module

Generate module from list of definitions.

terminator

Arguments

:: Integer
The 32-bit Integer to return.

-> Named Terminator
The return statement.

Generate a ret statement, returning a 32-bit Integer to the caller. While we use 64-bit integers everywhere else, our main function needs to return an int which usually is 32-bits even on 64-bit systems.

createGlobalString :: Lexeme -> Global

Generate global byte array (from a constant string).

generateConstants :: [AST] -> [Global]

Create constant strings/byte arrays for a module. TODO: Maybe rename these subfunctions?

getAllCons :: [AST] -> [Lexeme]

Get all Constants in a module.

createGlobalVariable :: Lexeme -> Global

Generate global variables for push and pop from and into variables.

generateVariables :: [AST] -> [Global]

Generate all global variable definitions for a module.

generateInstruction :: (Lexeme, String) -> Codegen [Named Instruction]

Generate an instruction for the underflow check command.

appendName :: [a] -> String -> [(a, String)]

Appends the function name to the lexemes/instructions.

The function name is only relevant for the Lambda instruction, because we use functionName!jumpLable as name for the Lambda functions.

generateBasicBlock :: ((Int, [Lexeme], Int), String) -> Codegen BasicBlock

Generate the instructions making up one basic block.

generateBasicBlocks :: [(Int, [Lexeme], Int)] -> String -> Codegen [BasicBlock]

Generate all basic blocks for a function.

generateFunction :: AST -> GlobalCodegen Definition

Generate a function definition from an AST.

At the beginning of each function, we create a block with lable entry. In this block we create the symbol table and jump then to l_1, the first regular block in each function. The first pattern-matching is for lambda functions

fresh :: Codegen Word

Create a new local variable (?).

generateFunctions :: [AST] -> GlobalCodegen [Definition]

Generate list of LLVM Definitions from list of ASTs

generateGlobalDefinition :: Integer -> Global -> Definition

Generate a global definition for a constant.

This is an unnamed, global constant, i. e. it has a numeric name like '@0'.

generateGlobalDefinitionVar :: Integer -> Global -> Definition

Generate a global definition for a variable name.

Such definitions are used to pass variable names to LLVM backend functions like 'pop_into()'.

process :: SemAna2InterCode -> InterCode2Backend

Entry point into module.