Lexer

Description

License : MIT

The lexer receives the output of the preprocessor -- a list of lists of strings, where each list represents a single function -- and turns it into a forest of function graphs. The forest is represented as a list of tuples; each tuple describes a function graph and contains the function name as a string as well as the function's graph itself.

A single function graph is a list of nodes. A node is a triple containing the following elements, in this order:

Node ID as an Integer. Starts with 1 for each function. * The Lexeme for this node. * The node ID of the follower node or 0 if there is no next node.

Note that for valid input, the only node with a follower ID of 0 can be a node containing the Finish lexeme. If any other node contains a follower ID of 0, this is an error (or, in Rail terms, a crash.

Documentation

process

Arguments

:: PreProc2Lexer
Preprocessor output (a list of lists of strings; i. e. a list of functions in their line representation).

-> Lexer2SynAna
A list of ASTs, each describing a single function.

Process preprocessor output into a list of function ASTs.

Raises errors on invalid input; see ErrorHandling for a list of error messages.

processfn

Arguments

:: Grid2D
The lines representing the function.

-> Graph
A graph of nodes representing the function. There may be more functions because of lambdas.

Process a single function.

funcname :: Grid2D -> Either String String

Get the name of the given function.

nodes

Arguments

:: Grid2D
Lines representing the function.

-> [[LexNode]]
Current graph representing the function. Initialize with [[(1, Start, 0, (0, 0, SE))]].

-> IP
Current instruction pointer. Initialize with start.

-> ([[LexNode]], IP)
Final graph for the function and the new instruction pointer.

Get the nodes for the given function.

handle

Arguments

:: Grid2D
Line representation of input function.

-> [[LexNode]]
Current list of nodes.

-> IP
Current instruction pointer.

-> ([[LexNode]], IP)
New node list and new instruction pointer.

Helper function for nodes: Handle the creation of the next PreLexNode for the current function.

update

Arguments

:: [[LexNode]]
The graph to operate on.

-> RelDirection

-> Int
ID of new follower to set for the first node in the list.

-> [[LexNode]]
Resulting graph.

Change the following node of the first (i. e. last, since the list is reversed) node in the graph.

merge :: [[LexNode]] -> [[LexNode]]

step

Arguments

:: Grid2D
Current function in its line representation.

-> IP
Current instruction pointer.

-> IP
New instruction pointer.

Move the instruction pointer a single step.

finalize

Arguments

:: [LexNode]
LexNodes to convert.

-> [LexNode]
Resulting list of LexNodes.

Brings it into right order