Preprocessor

Description

License : MIT

Stability : stable

Preprocessor gets the content of the input file and puts each rail function into a list of strings such that the first character of the first line is a dollar sign. Leading lines without a dollar sign in the input file are removed.

Documentation

process :: Input2PreProc -> PreProc2Lexer

convert :: [String] -> Grid2D

maximize :: [String] -> [String]

Makes the first line as long as max(max(lines),#lines) this is useful for the lexer to determine an upper bound for empty endless loops

notStartingWithDollar :: String -> Bool

Return False iff the first character is a dollar sign.

removeLines :: [String] -> ([String], Int)

Removes all leading strings from list until first string begins with a dollar sign.

groupFunctions :: ([String], Int) -> [([String], Int)]

Puts every rail function/program into its on grid such that the dollar sign is the first character in the first line.