SemanticalAnalysis

Description

License : MIT

There is no need for a semantical analysis at this time, therefore the function process equals the identity function

Documentation

process :: SynAna2SemAna -> SemAna2InterCode

splitlambda :: AST -> [AST]

validfollowers :: AST -> Bool

getids :: AST -> ([Int], [Int])

nomain :: [AST] -> Bool

duplicatefunctions :: [AST] -> Bool

allfunctions :: [AST] -> [String]

check :: AST -> AST

checknode :: (Int, [Lexeme], Int) -> (Int, [Lexeme], Int)

checklexeme :: Lexeme -> Lexeme