SyntacticalAnalysis

Description

License : MIT

Stability : stable

SyntacticalAnalysis receives output of Lexer and turns each rail function graph into a list of paths. Each path is a triple and contains a Path-ID, a list of lexemes and a Path-ID of the path that follows. Those lexemes are executed in order and sequentially.

Documentation

process :: Lexer2SynAna -> SynAna2SemAna

mapify :: [(Int, Lexeme, Int)] -> Map Int (Lexeme, Int) -> Map Int (Lexeme, Int)

pathes :: Map Int (Lexeme, Int) -> [Int] -> [(Int, [Lexeme], Int)]

generates all pathes of a graph

findPath :: Int -> Map Int (Lexeme, Int) -> [Int] -> (Int, [Lexeme], Int)

generates one path depending on initial node

fst' :: (a, b, c) -> a

fetch triple components

startNodes :: [LexNode] -> [Int]

generates a list of all nodes, which are needed to be initial nodes of path: 1 as functionstart; conditional jmp; indegree > 1